add parent class OAuthPRAWTest
