+++ tests/helper.py
@@ -106,6 +106,18 @@ class PRAWTest(unittest.TestCase):
         return urljoin(self.r.config.permalink_url, path)
 
 
+class OAuthPRAWTest(PRAWTest):
+    def betamax_init(self):
+        self.r.set_oauth_app_info(self.client_id,
+                                  self.client_secret,
+                                  self.redirect_uri)
+
+    def setUp(self):
+        self.configure()
+        self.r = Reddit(USER_AGENT, site_name='reddit_oauth_test',
+                        disable_update_check=True)
+
+
 Betamax.register_request_matcher(BodyMatcher)
 with Betamax.configure() as config:
     if os.getenv('TRAVIS'):
