+++ tests/test_subreddit.py
@@ -2,10 +2,10 @@
 
 from __future__ import print_function, unicode_literals
 import warnings
-from praw import Reddit, errors
+from praw import errors
 from praw.objects import Subreddit
 from six import text_type
-from .helper import PRAWTest, USER_AGENT, betamax
+from .helper import OAuthPRAWTest, PRAWTest, betamax
 
 
 class SubredditTest(PRAWTest):
@@ -268,17 +268,7 @@ class ModeratorSubredditTest(PRAWTest):
                         self.subreddit.get_wiki_contributors)
 
 
-class OAuthSubredditTest(PRAWTest):
-    def betamax_init(self):
-        self.r.set_oauth_app_info(self.client_id,
-                                  self.client_secret,
-                                  self.redirect_uri)
-
-    def setUp(self):
-        self.configure()
-        self.r = Reddit(USER_AGENT, site_name='reddit_oauth_test',
-                        disable_update_check=True)
-
+class OAuthSubredditTest(OAuthPRAWTest):
     @betamax()
     def test_add_remove_moderator_oauth(self):
         self.r.refresh_access_information(self.refresh_token['modothers'])
